+++ src/shapes/Object/InteractiveObject.spec.ts
@@ -4,7 +4,11 @@ import { Canvas } from '../../canvas/Canvas';
 import { FabricObject } from './FabricObject';
 import type { TOCoord } from './InteractiveObject';
 
-describe('Object', () => {
+describe('FabricObject', () => {
+  it('Interactive + BaseObject default values', () => {
+    const { controls, ...defaults } = FabricObject.getDefaults();
+    expect(defaults).toMatchSnapshot();
+  });
   describe('setCoords for objects inside group with rotation', () => {
     it('all corners are rotated as much as the object total angle', () => {
       const canvas = new Canvas();